refactor: keep KV capacity fixes in existing modules
