+++ vllm_ascend/patch/platform/patch_kv_cache_utils.py
@@ -22,11 +22,6 @@ from vllm.v1.kv_cache_interface import (
     get_kv_cache_spec_kind,
 )
 
-from vllm_ascend.core.kv_cache_capacity import (
-    replicated_draft_max_memory_usage_bytes,
-    replicated_draft_pool_bytes_per_block,
-)
-
 _KIMI_K3_TARGET_LAYER_PREFIX = "language_model.model.layers."
 _KIMI_K3_DRAFT_LAYER_PREFIX = "model.layers."
 _orig_resolve_kv_cache_block_sizes = vllm.v1.core.kv_cache_utils.resolve_kv_cache_block_sizes
@@ -451,8 +446,45 @@ def _is_deepseek_v4_groups(kv_cache_groups: list[KVCacheGroupSpec]) -> bool:
     return False
 
 
+def _replicated_draft_pool_bytes_per_block(groups: list[KVCacheGroupSpec]) -> int | None:
+    """Size the target/Mamba shared tensors plus independent replicated drafts.
+
+    The replicated-draft planner stores each target layer once and aliases
+    the corresponding Mamba layers into it. Draft tensors have larger pages
+    but do not add those pages to every target/Mamba layer tuple.
+    """
+    if len(groups) < 2 or not all(isinstance(group.kv_cache_spec, UniformTypeKVCacheSpecs) for group in groups):
+        return None
+    first = groups[0].kv_cache_spec.kv_cache_specs
+    drafts = [first[name] for name in groups[0].layer_names if getattr(first[name], "dcp_replication_size", 1) > 1]
+    targets = [first[name] for name in groups[0].layer_names if getattr(first[name], "dcp_replication_size", 1) == 1]
+    if not drafts or not targets or not all(isinstance(spec, FullAttentionSpec) for spec in (*targets, *drafts)):
+        return None
+    for group in groups[1:]:
+        specs = group.kv_cache_spec.kv_cache_specs
+        if len(group.layer_names) > len(targets):
+            return None
+        for name, target in zip(group.layer_names, targets):
+            spec = specs[name]
+            if not isinstance(spec, MambaSpec) or spec.page_size_bytes != target.page_size_bytes:
+                return None
+    return sum(spec.page_size_bytes for spec in (*targets, *drafts))
+
+
+def _replicated_draft_max_memory_usage_bytes(config: VllmConfig, groups: list[KVCacheGroupSpec]) -> int | None:
+    """Reserve enough shared pool slots for all groups of one max-length request."""
+    bytes_per_block = _replicated_draft_pool_bytes_per_block(groups)
+    if bytes_per_block is None:
+        return None
+    blocks = sum(
+        cdiv(group.kv_cache_spec.max_memory_usage_bytes(config), group.kv_cache_spec.page_size_bytes)
+        for group in groups
+    )
+    return blocks * bytes_per_block
+
+
 def _ascend_pool_bytes_per_block(kv_cache_groups: list[KVCacheGroupSpec]) -> int:
-    """Use the same DSV4 divisor as Ascend's shared-tuple planner.
+    """Match Ascend's replicated-draft and DSV4 pool layouts.
 
     vLLM #51718 re-plans ranks with more KV memory using
     ``min_num_blocks * _pool_bytes_per_block(groups)``. Its standardized
@@ -460,7 +492,7 @@ def _ascend_pool_bytes_per_block(kv_cache_groups: list[KVCacheGroupSpec]) -> int
     layout, so using the upstream value changes ``num_blocks`` during the
     re-plan and leaves ranks inconsistent.
     """
-    replicated_bytes = replicated_draft_pool_bytes_per_block(kv_cache_groups)
+    replicated_bytes = _replicated_draft_pool_bytes_per_block(kv_cache_groups)
     if replicated_bytes is not None:
         return replicated_bytes
     if not _is_deepseek_v4_groups(kv_cache_groups):
@@ -474,8 +506,8 @@ def _ascend_max_memory_usage_bytes_from_groups(
     vllm_config: VllmConfig,
     kv_cache_groups: list[KVCacheGroupSpec],
 ) -> int:
-    """Keep the pre-#51718 DSV4 admission formula for its shared tuples."""
-    replicated_bytes = replicated_draft_max_memory_usage_bytes(vllm_config, kv_cache_groups)
+    """Use layout-specific admission estimates for replicated drafts and DSV4."""
+    replicated_bytes = _replicated_draft_max_memory_usage_bytes(vllm_config, kv_cache_groups)
     if replicated_bytes is not None:
         return replicated_bytes
     if not _is_deepseek_v4_groups(kv_cache_groups):
